updated removing functionality for posts

+++ app/jobs/mass_change_status_for_posts_job.rb
@@ -6,6 +6,7 @@ class MassChangeStatusForPostsJob
     begin
       posts = FacebookPost.unscoped.where(mass_job_status: "to_be_#{new_status}")
       current_user = User.find(current_user_id)
+      ignore_array = FacebookPost::IGNORE_POST_FOR
     rescue
       raise "Records or User Not Found."
     end
@@ -17,8 +18,16 @@ class MassChangeStatusForPostsJob
       end
       posts.update_all(mass_job_status: :no_status)
     elsif new_status == 'ignored'
-      posts.each do |post|
-        post.destroy!
+      posts.find_each do |post|
+        if !post.added_by.present? and
+            ignore_array.include? post.blacklisted_by and
+            ignore_array.include? post.whitelisted_by and
+            ignore_array.include? post.greylisted_by  and
+            ignore_array.include? post.reported_to_facebook_by
+          post.destroy!
+        else
+          post.update(mass_job_status: :no_status)
+        end
       end
     end
